+++ model/card.go
@@ -5,13 +5,14 @@ import (
 	"gorm.io/gorm"
 )
 
+// TODO split card table and raw image data into separate tables and migrate database
 type Card struct {
 	gorm.Model
-	SpellID   uint64
+	SpellID   uint
 	Spell     Spell
 	Image     sql.RawBytes `json:"-"`
 	Version   string       // in what environment made card (tags set, version on StableDiffusion etc.)
-	PaintTime uint64       // seconds, how much paint took
+	PaintTime uint         // seconds, how much paint took
 }
 
 func (c Card) TableName() string {
