+++ model/spell.go
@@ -7,6 +7,6 @@ import "gorm.io/gorm"
 type Spell struct {
 	gorm.Model
 	Tags    string // additional tags to paint the picture (https://www.reddit.com/r/StableDiffusion/comments/y649yn/prompts_modifiers_to_get_midjourney_style_in/)
-	Seed    uint64 // specified seed (seed is from 0 to 10 000 000 000)
+	Seed    uint   // specified seed (seed is from 0 to 10 000 000 000)
 	Version string // in what environment made card (tags set, version on StableDiffusion etc.)
 }
